refactor: 공연 상세 정보 조회 내 등급 관련 정보 수정
