+++ src/main/java/com/culture/ticketing/show/application/dto/ShowDetailResponse.java
@@ -1,10 +1,8 @@
 package com.culture.ticketing.show.application.dto;
 
-import com.culture.ticketing.place.domain.Place;
+import com.culture.ticketing.show.domain.Place;
 import com.culture.ticketing.show.domain.Show;
 import com.culture.ticketing.show.round_performer.application.dto.RoundWithPerformersResponse;
-import com.culture.ticketing.show.show_floor.application.dto.ShowFloorGradeResponse;
-import com.culture.ticketing.show.show_seat.application.dto.ShowSeatGradeResponse;
 import lombok.Builder;
 import lombok.Getter;
 
@@ -15,25 +13,20 @@ public class ShowDetailResponse {
 
     private final ShowResponse show;
     private final List<RoundWithPerformersResponse> rounds;
-    private final List<ShowSeatGradeResponse> showSeatGrades;
-    private final List<ShowFloorGradeResponse> showFloorGrades;
+    private final List<ShowAreaGradeResponse> showAreaGrades;
 
     @Builder
-    public ShowDetailResponse(ShowResponse show, List<RoundWithPerformersResponse> rounds,
-                              List<ShowSeatGradeResponse> showSeatGrades, List<ShowFloorGradeResponse> showFloorGrades) {
+    public ShowDetailResponse(ShowResponse show, List<RoundWithPerformersResponse> rounds, List<ShowAreaGradeResponse> showAreaGrades) {
         this.show = show;
         this.rounds = rounds;
-        this.showSeatGrades = showSeatGrades;
-        this.showFloorGrades = showFloorGrades;
+        this.showAreaGrades = showAreaGrades;
     }
 
-    public static ShowDetailResponse from(Show show, Place place, List<RoundWithPerformersResponse> rounds,
-                                          List<ShowSeatGradeResponse> showSeatGrades, List<ShowFloorGradeResponse> showFloorGrades) {
+    public static ShowDetailResponse from(Show show, Place place, List<RoundWithPerformersResponse> rounds, List<ShowAreaGradeResponse> showAreaGrades) {
         return ShowDetailResponse.builder()
                 .show(ShowResponse.from(show, place))
                 .rounds(rounds)
-                .showSeatGrades(showSeatGrades)
-                .showFloorGrades(showFloorGrades)
+                .showAreaGrades(showAreaGrades)
                 .build();
     }
 }
